requirements.txt
pandas~=1.4.2
termcolor~=1.1.0
requests==2.28.1
python_telegram_bot~=13.3